requirements.txt
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
colorlog==4.0.2
Faker==2.0.2
greenlet==0.4.13
idna==2.8
pika==1.1.0
psycopg2cffi==2.8.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.3
python-logging-rabbitmq==2.0.0
readline==6.2.4.1
redis==3.3.8
requests==2.22.0
six==1.12.0
text-unidecode==1.3
urllib3==1.25.5
